requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
nltk
wordcloud
google-generativeai